refactor(app): batch5 — extract useSessionActions (after useConfirmSessionClose)
